extract boost code from chatbot
