+++ packages/client/src/views/chatbot/boost.ts
@@ -1,0 +1,64 @@
+import { Context, Language } from "decorator-shared/params";
+import loadExternalScript from "../../helpers/load-external-script";
+import { env, param } from "../../params";
+
+export const initBoost = (): Promise<{ show: () => void }> =>
+    loadScript().then(() => {
+        if (!window.boostInit) {
+            console.error("Boost init function not found!");
+            return { show: () => {} };
+        }
+
+        const boost = window.boostInit(
+            env("BOOST_ENV"),
+            buildBoostConfig({
+                context: param("context"),
+                language: param("language"),
+            }),
+        );
+
+        boost.chatPanel.addEventListener("setFilterValue", (event) => {
+            boost.chatPanel.setFilterValues(event.detail.filterValue);
+            if (event.detail.nextId) {
+                boost.chatPanel.triggerAction(event.detail.nextId);
+            }
+        });
+
+        return { show: boost.chatPanel.show };
+    });
+
+const buildBoostConfig = ({
+    context,
+    language,
+}: {
+    context: Context;
+    language: Language;
+}) => ({
+    chatPanel: {
+        settings: {
+            removeRememberedConversationOnChatPanelClose: true,
+            openTextLinksInNewTab: true,
+        },
+        styling: { buttons: { multiline: true } },
+        header: {
+            filters: {
+                filterValues:
+                    context === "arbeidsgiver"
+                        ? "arbeidsgiver"
+                        : language === "nn"
+                          ? "nynorsk"
+                          : "bokmal",
+            },
+        },
+    },
+});
+
+export const hasActiveConversation = (): boolean =>
+    null !==
+    (window.localStorage.getItem("boostai.conversation.id") ??
+        window.sessionStorage.getItem("boostai.conversation.id"));
+
+export const loadScript = () =>
+    loadExternalScript(
+        `https://${env("BOOST_ENV")}.boost.ai/chatPanel/chatPanel.js`,
+    );
